backend/requirements.txt
fastapi
uvicorn
jinja2

# 데이터베이스
sqlalchemy
psycopg[binary]
aiosqlite

# Google Gemini AI
google-generativeai

# 인증
passlib[bcrypt]
python-jose[cryptography]

# 환경 변수
python-dotenv

# 기타 유틸리티
python-multipart